update: architecture example, price single buy nft
