+++ mintbase_example/lib/modules/controllers/auth_controller.dart
@@ -3,7 +3,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutterchain/flutterchain_lib/constants/chains/near_blockchain_network_urls.dart';
 import 'package:flutterchain/flutterchain_lib/services/chains/near_blockchain_service.dart';
 import 'package:mintbase_example/consts/secure_storage_keys.dart';
-import 'package:mintbase_example/modules/models/models.dart';
+import 'package:mintbase_example/modules/models/auth_info_model/auth_info.dart';
 import 'package:rxdart/rxdart.dart';
 
 class AuthController extends Disposable {
