+++ mintbase_example/lib/modules/models/auth_info_model/auth_info.dart
@@ -2,8 +2,6 @@ import 'package:flutter/material.dart';
 
 enum NearNetworkType { testnet, mainnet }
 
-enum CategoryNFT { art, defi, fashion, membership, music, photography, ticket }
-
 @immutable
 class AuthInfo {
   final String accountId;
