+++ mintbase_example/lib/modules/models/category_nft_model/category_nft.dart
@@ -1,0 +1,1 @@
+enum CategoryNFT { art, defi, fashion, membership, music, photography, ticket }
