+++ mintbase_example/lib/modules/pages/auth/widgets/rolling_auction_nft.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutterchain/flutterchain_lib/services/chains/near_blockchain_service.dart';
-import 'package:mintbase_example/modules/pages/thems/thems.dart';
-import 'package:mintbase_example/modules/services/auth_controller.dart';
+import 'package:mintbase_example/thems/thems.dart';
+import 'package:mintbase_example/modules/controllers/auth_controller.dart';
 
 class RollingAuctionNft extends StatefulWidget {
   const RollingAuctionNft({super.key});
